Make Map fully responsive to page width
The length of the map was hard-coded in "vw" units so it needs a very
big browser window. Now the required browser window is smaller.

Also, the cell distortions are less pronounced now as the cell size are
calculated manually instead of using the CSS Grid "fr" unit, which tends
to be problematic in large grid.

Not solved: the addresses and the SideMenu needs shrinking to make the
required window size even smaller.